backend/requirements.txt
django
djangorestframework
django-cors-headers
scikit-learn
numpy
pdfminer.six
gunicorn
